chore: home page with Hello World
