+++ .eslintrc.json
@@ -1,3 +1,3 @@
 {
-    "extends": ["prettier"]
+    "extends": ["next/core-web-vitals"]
 }
