+++ package.json
@@ -14,6 +14,7 @@
         "format:check": "prettier . --check"
     },
     "dependencies": {
+        "@next/font": "14.2.3",
         "next": "14.2.3",
         "normalize.css": "8.0.1",
         "react": "18.2.0",
