+++ src/app/about.page.jsx
@@ -1,6 +1,6 @@
 'use strict'
 
-import styles from '@styles/About.module.css'
+import styles from '@/assets/css/About.module.css'
 import { Link } from 'react-router-dom'
 
 export const AboutPage = () => (
