+++ src/app/blog.page.jsx
@@ -1,6 +1,6 @@
 'use strict'
 
-import styles from '@styles/Blog.module.css'
+import styles from '@/assets/css/Blog.module.css'
 import { Link } from 'react-router-dom'
 
 export const BlogPage = () => (
