+++ src/app/page.tsx
@@ -1,3 +1,7 @@
 export default function Page() {
-    return <h1>Hello, World!</h1>
+    return (
+        <div className="container">
+            <h1>Hello, World!</h1>
+        </div>
+    )
 }
