+++ src/assets/css/Header.module.css
@@ -4,7 +4,7 @@
     position: sticky;
     top: 0;
 
-    > * {
+    & > * {
         display: flex;
         justify-content: space-between;
         align-items: center;
@@ -17,7 +17,7 @@
     border: 0;
     padding: 1.5rem 1rem;
 
-    svg {
+    & svg {
         fill: #fff;
         font-size: clamp(1rem, 1.5rem, 8vw);
     }
