+++ src/assets/css/Layout.module.css
@@ -1,0 +1,7 @@
+/* ----- Root Layout ----- */
+
+.Layout {
+    display: grid;
+    grid-template-rows: min-content 1fr min-content;
+    min-height: 100vh;
+}
