+++ src/assets/css/globals.css
@@ -11,17 +11,56 @@
     padding: 0;
 }
 
-html {
+html,
+body {
     box-sizing: inherit;
     font-size: 100%;
+    margin: 0;
     scroll-behavior: smooth;
 }
 
 body {
-    margin: 0;
+    font-family: Poppins, system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell,
+        'Fira Sans', 'Droid Sans', 'Open Sans', 'Helvetica Neue', sans-serif;
     min-height: 100vh;
     max-width: 100vw;
     min-width: 20rem;
-    overflow: auto;
     scrollbar-gutter: stable both-edges;
+}
+
+p {
+    margin-bottom: 0.75rem;
+}
+
+a {
+    font-weight: 500;
+    color: inherit;
+    text-decoration: none;
+}
+
+img {
+    max-width: 100%;
+    object-fit: cover;
+    vertical-align: middle;
+}
+
+li {
+    list-style-type: none;
+}
+
+.container {
+    max-width: clamp(calc(20rem - 2.5rem), 80vw, 95vw);
+    margin-left: auto;
+    margin-right: auto;
+}
+
+@media (prefers-color-scheme: dark) {
+    html {
+        color-scheme: dark;
+    }
+
+    body {
+        color: white;
+        background: black;
+    }
 }
