+++ src/components/LogoSquareSVG.tsx
