--- src/components/index.js
@@ -1,6 +1,0 @@
-'use strict'
-
-export { Header } from './Header'
-export { Footer } from './Footer'
-export { Logo } from './Logo'
-export { Navbar } from './Navbar'
