+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export * from './Header'
+export * from './Navbar'
+export * from './Footer'
+export * from './Logo'
+export * from './LogoCircleSVG'
+export * from './LogoSquareSVG'
