--- src/pagess/index.js
@@ -1,8 +1,0 @@
-'use strict'
-
-export { AboutPage } from './about.page'
-export { BlogPage } from './blog.page'
-export { ContactPage } from './contact.page'
-export { ErrorPage } from './error.page'
-export { HomePage } from './home.page'
-export { PortfolioPage } from './portfolio.page'
