--- src/partials/Head.tsx
@@ -1,13 +1,0 @@
-import 'normalize.css'
-import '@/assets/css/globals.css'
-
-export default function Head() {
-    return (
-        <head>
-            <meta charSet="UTF-8" />
-            <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-            <link rel="icon" href="/favicon.svg" type="image/svg+xml" />
-            <title>Sanders Gutiérrez</title>
-        </head>
-    )
-}
